Move FlowExtensionsTest
